+++ esp-hal/src/efuse/esp32c2/mod.rs
@@ -4,6 +4,7 @@ mod fields;
 pub use fields::*;
 
 /// Selects which ADC we are interested in the efuse calibration data for
+#[instability::unstable]
 pub enum AdcCalibUnit {
     /// Select efuse calibration data for ADC1
     ADC1,
@@ -11,6 +12,7 @@ pub enum AdcCalibUnit {
 
 impl super::Efuse {
     /// Get status of SPI boot encryption.
+    #[instability::unstable]
     pub fn flash_encryption() -> bool {
         !Self::read_field_le::<u8>(SPI_BOOT_CRYPT_CNT)
             .count_ones()
@@ -18,6 +20,7 @@ impl super::Efuse {
     }
 
     /// Get the multiplier for the timeout value of the RWDT STAGE 0 register.
+    #[instability::unstable]
     pub fn rwdt_multiplier() -> u8 {
         Self::read_field_le::<u8>(WDT_DELAY_SEL)
     }
@@ -25,6 +28,7 @@ impl super::Efuse {
     /// Get efuse block version
     ///
     /// see <https://github.com/espressif/esp-idf/blob/dc016f5987/components/hal/efuse_hal.c#L27-L30>
+    #[instability::unstable]
     pub fn block_version() -> (u8, u8) {
         // see <https://github.com/espressif/esp-idf/blob/dc016f5987/components/hal/esp32c2/include/hal/efuse_ll.h#L65-L73>
         // <https://github.com/espressif/esp-idf/blob/903af13e8/components/efuse/esp32c2/esp_efuse_table.csv#L90-L91>
@@ -37,6 +41,7 @@ impl super::Efuse {
     /// Get version of RTC calibration block
     ///
     /// see <https://github.com/espressif/esp-idf/blob/903af13e8/components/efuse/esp32c2/esp_efuse_rtc_calib.c#L14>
+    #[instability::unstable]
     pub fn rtc_calib_version() -> u8 {
         let (major, _minor) = Self::block_version();
         if major == 0 { 1 } else { 0 }
@@ -45,6 +50,7 @@ impl super::Efuse {
     /// Get ADC initial code for specified attenuation from efuse
     ///
     /// see <https://github.com/espressif/esp-idf/blob/903af13e8/components/efuse/esp32c2/esp_efuse_rtc_calib.c#L27>
+    #[instability::unstable]
     pub fn rtc_calib_init_code(_unit: AdcCalibUnit, atten: Attenuation) -> Option<u16> {
         let version = Self::rtc_calib_version();
 
@@ -74,6 +80,7 @@ impl super::Efuse {
     /// Get ADC reference point voltage for specified attenuation in millivolts
     ///
     /// see <https://github.com/espressif/esp-idf/blob/903af13e8/components/efuse/esp32c2/esp_efuse_rtc_calib.c#L65>
+    #[instability::unstable]
     pub fn rtc_calib_cal_mv(_unit: AdcCalibUnit, atten: Attenuation) -> u16 {
         match atten {
             Attenuation::_0dB => 400,
@@ -85,6 +92,7 @@ impl super::Efuse {
     /// Get ADC reference point digital code for specified attenuation
     ///
     /// see <https://github.com/espressif/esp-idf/blob/903af13e8/components/efuse/esp32c2/esp_efuse_rtc_calib.c#L65>
+    #[instability::unstable]
     pub fn rtc_calib_cal_code(_unit: AdcCalibUnit, atten: Attenuation) -> Option<u16> {
         let version = Self::rtc_calib_version();
 
@@ -116,11 +124,13 @@ impl super::Efuse {
     }
 
     /// Returns the major hardware revision
+    #[instability::unstable]
     pub fn major_chip_version() -> u8 {
         Self::read_field_le(WAFER_VERSION_MAJOR)
     }
 
     /// Returns the minor hardware revision
+    #[instability::unstable]
     pub fn minor_chip_version() -> u8 {
         Self::read_field_le(WAFER_VERSION_MINOR)
     }